AI_Server/requirements.txt
flask
flask-cors
numpy
pillow
torch
torchvision
faiss-cpu
scikit-learn
